chore(OrderCard.tsx): extract status badge as a separate component. chore(StatusBadge.tsx): add status badge component.
